function version of ransac
